Remove separate warm/cold VDDK importers.
