+++ pkg/importer/BUILD.bazel
@@ -12,9 +12,7 @@ go_library(
         "transport.go",
         "upload-datasource.go",
         "util.go",
-        "vddk-common.go",
         "vddk-datasource.go",
-        "vddk-incremental-datasource.go",
     ],
     importpath = "kubevirt.io/containerized-data-importer/pkg/importer",
     visibility = ["//visibility:public"],
